1666: Fix multiple issues and use PlainDate instead of epochs
