+++ administration/src/cards/CardBlueprint.test.ts
@@ -1,8 +1,19 @@
 import { BavariaCardType } from '../generated/card_pb'
 import { Region } from '../generated/graphql'
 import PlainDate from '../util/PlainDate'
-import { generateCardInfo, initializeCardBlueprint } from './CardBlueprint'
+import {
+  generateCardInfo,
+  getValueByCSVHeader,
+  initializeCardBlueprint,
+  initializeCardFromCSV,
+  isValid,
+  isValueValid,
+} from './CardBlueprint'
 import BavariaCardTypeExtension, { BAVARIA_CARD_TYPE_EXTENSION_NAME } from './extensions/BavariaCardTypeExtension'
+import BirthdayExtension, { BIRTHDAY_EXTENSION_NAME } from './extensions/BirthdayExtension'
+import KoblenzReferenceNumberExtension, {
+  KOBLENZ_REFERENCE_NUMBER_EXTENSION_NAME,
+} from './extensions/KoblenzReferenceNumberExtension'
 import RegionExtension, { REGION_EXTENSION_NAME } from './extensions/RegionExtension'
 
 jest.useFakeTimers({ now: new Date('2020-01-01') })
@@ -28,7 +39,7 @@ describe('CardBlueprint', () => {
     const card = initializeCardBlueprint(cardConfig, region, { fullName: 'Thea Test' })
 
     expect(card.fullName).toBe('Thea Test')
-    expect(card.expirationDate).toEqual(PlainDate.from('2023-01-01').toDaysSinceEpoch())
+    expect(card.expirationDate).toEqual(PlainDate.from('2023-01-01'))
     expect(card.extensions[BAVARIA_CARD_TYPE_EXTENSION_NAME]).toBe('Standard')
     expect(card.extensions[REGION_EXTENSION_NAME]).toEqual(region.id)
   })
@@ -36,7 +47,7 @@ describe('CardBlueprint', () => {
   it('should generate CardInfo even with invalid expiration date', () => {
     const card = initializeCardBlueprint(cardConfig, region, {
       fullName: '',
-      expirationDate: PlainDate.from('1900-01-01').toDaysSinceEpoch(),
+      expirationDate: PlainDate.from('1900-01-01'),
     })
     expect(generateCardInfo(card).toJson({ enumAsInteger: true })).toEqual({
       fullName: '',
@@ -49,6 +60,118 @@ describe('CardBlueprint', () => {
           cardType: BavariaCardType.STANDARD,
         },
       },
+    })
+  })
+
+  describe('csv', () => {
+    const region: Region = {
+      id: 0,
+      name: 'augsburg',
+      prefix: 'a',
+      activatedForApplication: true,
+      activatedForCardConfirmationMail: true,
+    }
+
+    const cardConfig = {
+      defaultValidity: { years: 3 },
+      nameColumnName: 'Name',
+      expiryColumnName: 'Ablaufdatum',
+      extensionColumnNames: ['Kartentyp', null],
+      extensions: [BavariaCardTypeExtension, RegionExtension],
+    }
+
+    it('should correctly initialize CSVCard', () => {
+      const card = initializeCardFromCSV(cardConfig, [], [], region)
+
+      expect(card.fullName).toBe('')
+      expect(card.expirationDate).toBeNull()
+      expect(card.extensions[BAVARIA_CARD_TYPE_EXTENSION_NAME]).toBeUndefined()
+      expect(card.extensions[REGION_EXTENSION_NAME]).toBe(0)
+    })
+
+    it('should correctly set and get value', () => {
+      const dateString = '03.04.2022'
+      const date = PlainDate.fromCustomFormat(dateString)
+      const line = ['Thea Test', dateString, 'Goldkarte']
+      const headers = ['Name', 'Ablaufdatum', 'Kartentyp']
+      const card = initializeCardFromCSV(cardConfig, line, headers, region)
+
+      expect(card.fullName).toBe('Thea Test')
+      expect(card.expirationDate).toEqual(date)
+      expect(card.extensions[BAVARIA_CARD_TYPE_EXTENSION_NAME]).toBe('Goldkarte')
+
+      expect(isValueValid(card, cardConfig, 'Name')).toBeTruthy()
+      expect(isValueValid(card, cardConfig, 'Ablaufdatum')).toBeTruthy()
+      expect(isValueValid(card, cardConfig, 'Kartentyp')).toBeTruthy()
+      expect(isValid(card)).toBeTruthy()
+
+      expect(getValueByCSVHeader(card, cardConfig, 'Name')).toBe('Thea Test')
+      expect(getValueByCSVHeader(card, cardConfig, 'Ablaufdatum')).toBe(date.toString())
+      expect(getValueByCSVHeader(card, cardConfig, 'Kartentyp')).toBe('Goldkarte')
+    })
+
+    it('should not modify value for invalid header', () => {
+      const line = ['1']
+      const headers = ['Region']
+      const card = initializeCardFromCSV(cardConfig, line, headers, region)
+
+      expect(card.fullName).toBe('')
+
+      expect(getValueByCSVHeader(card, cardConfig, 'Region')).toBeNull()
+    })
+
+    it('should correctly identify invalid values', () => {
+      const line = ['2012/03/02', 'Graublau']
+      const headers = ['Ablaufdatum', 'Kartentyp']
+      const card = initializeCardFromCSV(cardConfig, line, headers, region)
+
+      expect(card.fullName).toBe('')
+      expect(getValueByCSVHeader(card, cardConfig, 'Name')).toBe('')
+      expect(card.expirationDate).toBeNull()
+      expect(getValueByCSVHeader(card, cardConfig, 'Ablaufdatum')).toBeNull()
+      expect(card.extensions[BAVARIA_CARD_TYPE_EXTENSION_NAME]).toBeUndefined()
+      expect(getValueByCSVHeader(card, cardConfig, 'Kartentyp')).toBeNull()
+
+      expect(isValueValid(card, cardConfig, 'Name')).toBeFalsy()
+      expect(isValueValid(card, cardConfig, 'Ablaufdatum')).toBeFalsy()
+      expect(isValueValid(card, cardConfig, 'Kartentyp')).toBeFalsy()
+      expect(isValid(card)).toBeFalsy()
+    })
+  })
+
+  describe('self service', () => {
+    const cardConfig = {
+      defaultValidity: { years: 3 },
+      nameColumnName: 'Name',
+      expiryColumnName: 'Ablaufdatum',
+      extensionColumnNames: ['Geburtsdatum', 'Referenznummer'],
+      extensions: [BirthdayExtension, KoblenzReferenceNumberExtension],
+    }
+    const expirationDate = PlainDate.fromLocalDate(new Date()).add(cardConfig.defaultValidity)
+
+    it('should correctly initialize CardBlueprint', () => {
+      const card = initializeCardBlueprint(cardConfig, undefined, { fullName: 'Karla Koblenz' })
+
+      expect(card.fullName).toBe('Karla Koblenz')
+      expect(card.expirationDate).toEqual(expirationDate)
+      expect(card.extensions[BIRTHDAY_EXTENSION_NAME]).toBe(BirthdayExtension.getInitialState().birthday)
+      expect(card.extensions[KOBLENZ_REFERENCE_NUMBER_EXTENSION_NAME]).toBe(
+        KoblenzReferenceNumberExtension.getInitialState().koblenzReferenceNumber
+      )
+    })
+
+    it('should generate CardInfo', () => {
+      const card = initializeCardBlueprint(cardConfig, undefined, { fullName: 'Karla Koblenz' })
+      expect(generateCardInfo(card).toJson()).toEqual({
+        fullName: 'Karla Koblenz',
+        expirationDay: expirationDate.toDaysSinceEpoch(),
+        extensions: {
+          extensionBirthday: {},
+          extensionKoblenzReferenceNumber: {
+            referenceNumber: '',
+          },
+        },
+      })
     })
   })
 })
