+++ administration/src/cards/extensions/StartDayExtension.tsx
@@ -6,14 +6,12 @@ import PlainDate from '../../util/PlainDate'
 import { Extension, ExtensionComponentProps } from './extensions'
 
 export const START_DAY_EXTENSION_NAME = 'startDay'
-export type StartDayExtensionState = { [START_DAY_EXTENSION_NAME]: number }
+export type StartDayExtensionState = { [START_DAY_EXTENSION_NAME]: PlainDate }
 
 // Some minimum start day after 1970 is necessary, as we use an uint32 in the protobuf.
 const minStartDay = new PlainDate(2020, 1, 1)
 
 const StartDayForm = ({ value, setValue, isValid }: ExtensionComponentProps<StartDayExtensionState>): ReactElement => {
-  const startDayDate = PlainDate.fromDaysSinceEpoch(value.startDay)
-
   return (
     <FormGroup label='Startdatum'>
       <TextField
@@ -22,14 +20,14 @@ const StartDayForm = ({ value, setValue, isValid }: ExtensionComponentProps<Star
         required
         size='small'
         error={!isValid}
-        value={startDayDate.toString()}
+        value={value.startDay.toString()}
         sx={{ '& input[value=""]:not(:focus)': { color: 'transparent' }, '& fieldset': { borderRadius: 0 } }}
         inputProps={{
           min: minStartDay.toString(),
           style: { fontSize: 14, padding: '6px 10px' },
         }}
         onChange={event => {
-          const date = PlainDate.safeEpochsFrom(event.target.value)
+          const date = PlainDate.safeFrom(event.target.value)
           if (date !== null) {
             setValue({ startDay: date })
           }
@@ -39,25 +37,24 @@ const StartDayForm = ({ value, setValue, isValid }: ExtensionComponentProps<Star
   )
 }
 
-const isStartDayValid = ({ startDay }: StartDayExtensionState): boolean =>
-  PlainDate.fromDaysSinceEpoch(startDay).isBefore(minStartDay)
+const isStartDayValid = ({ startDay }: StartDayExtensionState): boolean => startDay.isAfter(minStartDay)
 
 const StartDayExtension: Extension<StartDayExtensionState> = {
   name: START_DAY_EXTENSION_NAME,
   Component: StartDayForm,
-  getInitialState: () => ({ startDay: PlainDate.fromLocalDate(new Date()).toDaysSinceEpoch() }),
+  getInitialState: () => ({ startDay: PlainDate.fromLocalDate(new Date()) }),
   causesInfiniteLifetime: () => false,
   getProtobufData: (state: StartDayExtensionState) => ({
     extensionStartDay: {
-      startDay: isStartDayValid(state) ? state.startDay : minStartDay.toDaysSinceEpoch(),
+      startDay: isStartDayValid(state) ? state.startDay.toDaysSinceEpoch() : minStartDay.toDaysSinceEpoch(),
     },
   }),
   isValid: isStartDayValid,
   fromString: (value: string) => {
-    const startDay = PlainDate.safeEpochsFromCustomFormat(value)
+    const startDay = PlainDate.safeFromCustomFormat(value)
     return startDay === null ? null : { startDay }
   },
-  toString: ({ startDay }: StartDayExtensionState) => PlainDate.fromDaysSinceEpoch(startDay).format(),
+  toString: ({ startDay }: StartDayExtensionState) => startDay.format(),
 }
 
 export default StartDayExtension
